requirements.txt
pyqt5
pysdl2
numpy
scipy
pyaudio
pyqtgraph
